Agora o render é separado em métodos com mais foco castRay melhorado Stub de shade

+++ RayTracerPG/Scene.cpp
@@ -7,7 +7,7 @@ void Scene::add(Object* object)
 
 int Scene::getNumberObjects()
 {
-	return this->objects.size();
+	return (int) this->objects.size();
 }
 
 Object* Scene::getObject(int object)
